pi/requirements.txt
RPi.GPIO
hx711
requests
python-socketio[client]
